requirements.txt
allure-pytest==2.13.5
pytest~=8.2.2
python-dotenv==1.0.1
selenium~=4.23.0
urllib3==2.2.1
requests~=2.32.3
Faker~=26.0.0